Draft of WINAPI related experiment
This module contains code aimed at locating a certain data structure in memory (in this case, an in-memory SQLite db is targeted). Here, the goal is to locate an in-memory SQLite db header, parse it, determine db size, and dump it to a file. Routines are drafted for searching both own memory address space and another process's memory.